Moves error keys into own utility file.
